Changes folder structure
