+++ frontend/src/App.jsx
@@ -2,13 +2,13 @@ import {Routes, Route } from "react-router-dom";
 import Layout from "./pages/Layout.jsx";
 import LandingPage from "./pages/LandingPage.jsx";
 import NotFound from "./pages/NotFound.jsx";
-import BlogPage from "./pages/BlogPage.jsx";
+import BlogPage from "./pages/blog/BlogPage.jsx";
 import AdmMenu from "./pages/adm/AdmMenu.jsx";
 import EditContact from "./pages/adm/EditContact.jsx";
-import BlogPost from "./pages/BlogPost.jsx";
-import EditBlogPost from "./pages/EditBlogPost.jsx";
+import BlogPost from "./pages/blog/BlogPost.jsx";
+import EditBlogPost from "./pages/blog/EditBlogPost.jsx";
 import RegisterAdm from "./pages/adm/RegisterAdm.jsx";
-import BazarPage from "./pages/BazarPage.jsx";
+import BazarPage from "./pages/blog/BazarPage.jsx";
 import Login from "./pages/Login.jsx";
 import axios from "axios";
 import ProtectedRoute from "./components/ProtectedRoute.jsx";
